requirements.txt
blinker==1.9.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
Flask==3.1.1
Flask-Limiter==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-PyMongo==2.3.0
Flask-WTF==1.2.2
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mongomock==4.3.0
openpyxl==3.1.5
packaging==25.0
pluggy==1.6.0
pycparser==2.22
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.4
text-unidecode==1.3
tomli==2.2.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
ruff==0.1.11